- Moved TopicRequest and TopicResponse to data - Added empty Topic constructor
