+++ Database/src/test/java/testers/listeners/TopicListenerTester.java
@@ -6,7 +6,6 @@ import java.util.concurrent.LinkedBlockingQueue;
 
 import testers.AbstractTester;
 import data.*;
-import listeners.*;
 import repositories.*;
 
 import org.junit.After;
