+++ Frontend/src/main/java/PIM/LoginController.java
@@ -15,7 +15,6 @@ import java.lang.Thread;
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.net.*;
-import listeners.*;
 import data.*;
 import org.springframework.beans.factory.annotation.*;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
