+++ data/.gitignore
@@ -1,3 +1,4 @@
 // ignore build folder
 build/*
 javadoc/
+.gradle/
