+++ data/compile.bash
@@ -1,7 +1,8 @@
 #!/bin/bash
 mkdir -p build/data/
+mkdir -p build/libs/
 cd src/data/
-javac -cp ".:../libs/org.springframework.data.core_1.3.1.RELEASE.jar" -d ../../build *.java
+javac -cp "../libs/*:." -d ../../build *.java
 cd ../../build/
-jar cf "data.jar" */* 
+jar cf "data.jar" */*
 mv data.jar ../data.jar
